+++ NEWS.md
@@ -1,3 +1,4 @@
 # LinCDE 0.2.1
 
+* Updated the predict function of LinCDE to allow more flexible inputs.
 * Added a `NEWS.md` file to track changes to the package.